Hw04 auth

+++ models/contacts.js
@@ -1,56 +1,59 @@
 import Contact from "../services/schemas/contacts.js";
 
-export const listContacts = async () => {
+export const listContacts = async (userId) => {
   try {
-    return await Contact.find();
+    return await Contact.find({ owner: userId });
   } catch (err) {
     console.error("Error getting contacts list: ", err);
     throw err;
   }
 };
 
-export const getContactById = async (contactId) => {
+export const getContactById = async (userId, contactId) => {
   try {
-    return await Contact.findOne({ _id: contactId });
+    return await Contact.findOne({ owner: userId, _id: contactId });
   } catch (err) {
     console.error(`Error getting contact with id: ${contactId}`, err);
     throw err;
   }
 };
 
-export const removeContact = async (contactId) => {
+export const removeContact = async (userId, contactId) => {
   try {
-    return await Contact.findByIdAndRemove({ _id: contactId });
+    return await Contact.findByIdAndRemove({ owner: userId, _id: contactId });
   } catch (err) {
     console.error(`Error removing contact with id: ${contactId}`, err);
     throw err;
   }
 };
 
-export const addContact = async (body) => {
+export const addContact = async (body, userId) => {
   try {
-    return await Contact.create(body);
+    const contacts = { ...body, owner: userId };
+    return await Contact.create(contacts);
   } catch (err) {
     console.error(`Error adding contact `, err);
     throw err;
   }
 };
 
-export const updateContact = async (contactId, body) => {
+export const updateContact = async (contactId, body, userId) => {
   try {
-    return await Contact.findByIdAndUpdate({ _id: contactId }, body, {
-      new: true,
-    });
+    return await Contact.findByIdAndUpdate(
+      { owner: userId, _id: contactId },
+      body,
+      { new: true }
+    );
   } catch (err) {
     console.error(`Error updating contact: `, err);
     throw err;
   }
 };
 
-export const updatedStatusContact = async (contactId, favorite) => {
+export const updatedStatusContact = async (userId, contactId, favorite) => {
   try {
     return await Contact.findByIdAndUpdate(
-      { _id: contactId },
+      { owner: userId, _id: contactId },
       { $set: { favorite: favorite } },
       { new: true }
     );
